+++ src/components/BurgerIngredients/BurgerIngredient.tsx
@@ -3,8 +3,12 @@ import classNames from "classnames";
 import { Tabs } from "../Tabs/Tabs";
 import React, { FC, useCallback, useEffect, useRef, useState } from "react";
 import { IngredientsHeap } from "../IngredientsHeap/IngredientsHeap";
-import { IBurgerIngredient, IIngredient } from "../../types/types";
+import { IIngredient } from "../../types/types";
 import { useAppDispatch, useAppSelector } from "../../services/store/store";
+
+interface IBurgerIngredient {
+  items?: IIngredient[];
+}
 
 export const BurgerIngredient: FC<IBurgerIngredient> = ({ items = [] }) => {
   const dispatch = useAppDispatch();
