+++ src/components/NavItem/NavItem.tsx
@@ -1,7 +1,12 @@
 import styles from "./styles.module.css";
 import classNames from "classnames";
 import { FC } from "react";
-import { INavItem } from "../../types/types";
+
+export interface INavItem {
+  text: string;
+  icon: JSX.Element;
+  isActive: boolean;
+}
 
 export const NavItem: FC<INavItem> = (props) => {
   return (
