+++ src/pages/profile/Profile.comonent.tsx
@@ -40,6 +40,7 @@ export const Profile = () => {
 
   const onExit = () => {
     dispatch(logoutApp());
+    console.log("!!!!!!!!!!!!!!!!!!!!!");
     navigate("/login");
   };
