+++ src/services/reducers/stateAppBehavior.ts
@@ -1,6 +1,8 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 
-import { IStateAppBehavior } from "../../types/types";
+interface IStateAppBehavior {
+  headerActive?: string;
+}
 
 const initialState: IStateAppBehavior = {
   headerActive: "",
